Move Axis components to same directory
