+++ src/client/components/Reuse/d3/Axes/Axes.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import XAxis from '../XAxis'
-import YAxis from '../YAxis'
+import XAxis from './XAxis'
+import YAxis from './YAxis'
 
 export interface AxisProperties {
     label: string
