+++ src/client/components/Reuse/d3/Axes/XAxis/XAxis.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react'
 import * as d3 from 'd3'
 
-import { AxisProperties } from '../Axes'
+import { AxisProperties } from '..'
 
 interface XAxisProps {
     domain: [number, number]
